# swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/javascript": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/style": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/inserir": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/categorias": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/deletarCategorias/{categoria_id}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "categoria_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/inserirCategorias": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/controllerFormCategoria": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "categoria_id": {
                  "example": "any"
                },
                "nome_categoria": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/selectCategorias/{id?}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id?",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/updateCategorias/{categoria_id}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "categoria_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/controllerUpdateCategorias": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome_categoria": {
                  "example": "any"
                },
                "categoria_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/cliente": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/inserirCliente": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/controllerFormCliente": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cliente_id": {
                  "example": "any"
                },
                "endereco_cliente": {
                  "example": "any"
                },
                "nome_cliente": {
                  "example": "any"
                },
                "cel_cliente": {
                  "example": "any"
                },
                "email_cliente": {
                  "example": "any"
                },
                "senha_cliente": {
                  "example": "any"
                },
                "datanasc_cliente": {
                  "example": "any"
                },
                "rg_cliente": {
                  "example": "any"
                },
                "cpf_cliente": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/selectCliente/{id?}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id?",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/updateCliente/{cliente_id}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/controllerUpdateCliente": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome_cliente": {
                  "example": "any"
                },
                "cliente_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/pagamento": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/inserirPagamento": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/controllerFormPagamento": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id_pagamento": {
                  "example": "any"
                },
                "forma_pagamento": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/selectPagamento/{id?}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id?",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/produto": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/inserirProduto": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/controllerFormProduto": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "produto_id": {
                  "example": "any"
                },
                "nome_produto": {
                  "example": "any"
                },
                "preco_produto": {
                  "example": "any"
                },
                "categoria_id": {
                  "example": "any"
                },
                "vendedor_id": {
                  "example": "any"
                },
                "qtd_estoque": {
                  "example": "any"
                },
                "forma_pagamento": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/selectProduto/{id?}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id?",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/venda": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/venda_do_produto": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/vendedor": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/controllerForm": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {}
}